requirements.txt
requests
python-dotenv
windows-curses